+++ gym_stag_hunt/src/games/harvest_game.py
@@ -1,6 +1,6 @@
 from random import uniform
 
-from numpy import zeros
+from numpy import zeros, uint8
 
 from gym_stag_hunt.src.games.abstract_grid_game import AbstractGridGame
 from gym_stag_hunt.src.utils import overlaps_entity, place_entity_in_unoccupied_cell
@@ -176,7 +176,7 @@ class Harvest(AbstractGridGame):
         maturity_flags = self.MATURITY_FLAGS
         plant_data = []
         for idx, element in enumerate(self.PLANTS):
-            new_entry = zeros(3, dtype=int)
+            new_entry = zeros(3, dtype=uint8)
             new_entry[0], new_entry[1], new_entry[2] = element[0], element[1], maturity_flags[idx]
             plant_data.append(new_entry)
         shipback = shipback, plant_data
